Test(Problem): 이전 커밋에 대해 미변경된 테스트 수정
